requirements.txt
fastapi
pydantic
pymongo
google-cloud-storage
bcrypt
pyjwt
python-dotenv
secure-smtplib
pydantic[email]
python-multipart
uvicorn
razorpay
pycryptodome
twilio
